requirements.txt
# Multi Vanna Dependencies

# Vanna AI
vanna[azureopenai,chromadb,postgres]>=2.0.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Database
psycopg2-binary>=2.9.9

# HTTP Client
httpx>=0.25.0

# AWS SDK (for ECS version)
boto3>=1.34.0
botocore>=1.34.0

# Environment
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
pydantic>=2.0.0

# LangSmith (optional, for tracing)
langsmith>=0.1.0
